Added the repository component and conncted to DB
Added the AuthRepository class to represent the repository layer of
the application, and modified the Article class to be represented
as JPA Entity to be persisted in the DB as a table. Also, removed
the field of in memory created articles from AuthService class and
replaced that field with the repository injected instance.

@MasaadAlotaibi
masaadbal@gmail.com
+966 53 533 8844