Resolved the issue of using JWT token in requests
The issue was that the requests to any endpoint not publicly
acessible is forbidden, that is requests to any endpoint other than
the endpoint POST"/login", and the bug was in parseClaimsJwt, as it
is not the right method to call, and I should have called the method
parseClaimsJws (notice the method names endings - the latter ends with
letter 's' while the former ends with 't'). The issue exactly was in
getting the claims in getClaims(...) method that is dependent upon
by isTokenExpired(...) which checks for expiry of sent token on each
request, and also dependent upon by getUsernameFromToken(...) which
extracts the username from the token and later validate it with
actual username. With all of this, any request using this token
will not be valiadated appropriately, hence Forbidden access issue
will be thrown.

@masaaadotaibi
masaadbal@gmail.com